requirements.txt
streamlit==1.28.0
pymupdf==1.23.8
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
python-dotenv==1.0.0
pypdf==3.17.0
faiss-cpu==1.7.4
openai==0.28.0
pandas==2.0.3
sentence-transformers==2.2.2
tiktoken==0.5.1
# Remove torch to avoid DLL issues, use CPU version
--index-url https://download.pytorch.org/whl/cpu
torch==2.0.1
torchvision==0.15.2
# Optional for tests
reportlab==4.0.4